+++ lib/modules/manager/bun/extract.spec.ts
@@ -1,6 +1,6 @@
-import { extractAllPackageFiles } from '.';
+import { extractAllPackageFiles } from './extract';
 
-describe('modules/manager/bun/extract/index', () => {
+describe('modules/manager/bun/extract', () => {
   describe('extractAllPackageFiles()', () => {
     it('ignores non-bun files', async () => {
       expect(await extractAllPackageFiles({}, ['package.json'])).toEqual([]);
