+++ lib/modules/manager/bun/extract.ts
